:sparkles: Adding 'route guard'
Now we have added a Route Guard to our Chat page, so if the user didnt fill the name form and try to enter in our chat adding '/ chat' to the URL, the user will be redirect to the 'Welcome' page. If the form is valid, the Route Guard will be responsible for forwarding the user to the next page.